+++ Assets/Core/Runtime/DataBind/ImageArrayBind.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine.UI;
+
+namespace KKSFramework.DataBind
+{
+    public class ImageArrayBind : BindableComponents<Image>
+    {
+    }
+}