+++ Assets/Core/Runtime/DataBind/TextArrayBind.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine.UI;
+
+namespace KKSFramework.DataBind
+{
+    public class TextArrayBind : BindableComponents<Text>
+    {
+    }
+}